# CarCompany version 2 (ASP Assignment 2)/swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "host": "localhost:5001",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json-patch+json",
    "application/json",
    "text/json",
    "application/*+json"
  ],
  "paths": {
...
    "/api/Skills": {
      "get": {
        "tags": [
          "Skills"
        ],
        "operationId": "Skills_GetSkills",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Skill"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Skills"
        ],
        "operationId": "Skills_PostSkill",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "skill",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Skill"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/api/Skills/{id}": {
      "get": {
        "tags": [
          "Skills"
        ],
        "operationId": "Skills_GetSkill",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Skills"
        ],
        "operationId": "Skills_PutSkill",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "skill",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Skill"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Skills"
        ],
        "operationId": "Skills_DeleteSkill",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/api/Users": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_GetUsers",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_PostUser",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/api/Users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_GetUser",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_PutUser",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "x-nullable": true
          },
          {
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_DeleteUser",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Car": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "id",
        "range",
        "mileage"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "make": {
          "type": "string"
        },
        "model": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "range": {
          "type": "number",
          "format": "double"
        },
        "mileage": {
          "type": "number",
          "format": "double"
        },
        "companyId": {
          "type": "integer",
          "format": "int32"
        },
        "userId": {
          "type": "string"
        }
      }
    },
...
    "Project": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "id",
        "projectType"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "projectType": {
          "$ref": "#/definitions/EProjectType"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "imageUrl": {
          "type": "string"
        },
        "startDate": {
          "type": "string",
          "format": "date-time"
        },
        "endDate": {
          "type": "string",
          "format": "date-time"
        },
        "userId": {
          "type": "string"
        },
        "skills": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Skill"
          }
        }
      }
    },
    "EProjectType": {
      "type": "integer",
      "description": "",
      "x-enumNames": [
        "FullTime",
        "PartTime",
        "Freelance"
      ],
      "enum": [
        0,
        1,
        2
      ]
    },
    "Skill": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "companyId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "User": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "age",
        "lastActivityDate"
      ],
      "properties": {
        "userName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "age": {
          "type": "integer",
          "format": "int32"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "lastActivityDate": {
          "type": "string",
          "format": "date-time"
        },
        "projects": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Project"
          }
        },
        "cars": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Car"
          }
        }
      }
    }
  }
}